SONAR-7353 Move RuleType to API
