+++ server/sonar-server/src/main/java/org/sonar/server/computation/issue/IssueCounter.java
@@ -28,8 +28,8 @@ import java.util.Map;
 import javax.annotation.Nullable;
 import org.sonar.api.issue.Issue;
 import org.sonar.api.measures.CoreMetrics;
+import org.sonar.api.rules.RuleType;
 import org.sonar.core.issue.DefaultIssue;
-import org.sonar.core.rule.RuleType;
 import org.sonar.server.computation.component.Component;
 import org.sonar.server.computation.measure.Measure;
 import org.sonar.server.computation.measure.MeasureRepository;
