+++ server/sonar-server/src/main/java/org/sonar/server/computation/issue/RuleImpl.java
@@ -26,9 +26,9 @@ import javax.annotation.Nullable;
 import javax.annotation.concurrent.Immutable;
 import org.sonar.api.rule.RuleKey;
 import org.sonar.api.rule.RuleStatus;
+import org.sonar.api.rules.RuleType;
 import org.sonar.api.server.debt.DebtRemediationFunction;
 import org.sonar.api.server.debt.internal.DefaultDebtRemediationFunction;
-import org.sonar.core.rule.RuleType;
 import org.sonar.db.rule.RuleDto;
 
 import static com.google.common.collect.Sets.union;
