+++ server/sonar-server/src/main/java/org/sonar/server/rule/index/RuleDoc.java
@@ -28,7 +28,7 @@ import javax.annotation.Nullable;
 import org.apache.commons.lang.builder.ReflectionToStringBuilder;
 import org.sonar.api.rule.RuleKey;
 import org.sonar.api.rule.RuleStatus;
-import org.sonar.core.rule.RuleType;
+import org.sonar.api.rules.RuleType;
 import org.sonar.server.search.BaseDoc;
 
 /**
