+++ server/sonar-server/src/main/java/org/sonar/server/rule/index/RuleQuery.java
@@ -25,7 +25,7 @@ import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
 import org.sonar.api.rule.RuleStatus;
 import org.sonar.api.rule.Severity;
-import org.sonar.core.rule.RuleType;
+import org.sonar.api.rules.RuleType;
 
 import static java.util.Arrays.asList;
