+++ server/sonar-server/src/test/java/org/sonar/server/computation/issue/IssueAssignerTest.java
@@ -20,10 +20,10 @@
 package org.sonar.server.computation.issue;
 
 import org.junit.Test;
+import org.sonar.api.rules.RuleType;
 import org.sonar.api.utils.log.LogTester;
 import org.sonar.api.utils.log.LoggerLevel;
 import org.sonar.core.issue.DefaultIssue;
-import org.sonar.core.rule.RuleType;
 import org.sonar.server.computation.analysis.AnalysisMetadataHolderRule;
 import org.sonar.server.computation.component.Component;
 import org.sonar.server.computation.scm.Changeset;
