+++ server/sonar-server/src/test/java/org/sonar/server/computation/issue/UpdateConflictResolverTest.java
@@ -24,9 +24,9 @@ import org.mockito.ArgumentCaptor;
 import org.sonar.api.issue.Issue;
 import org.sonar.api.rule.RuleKey;
 import org.sonar.api.rule.Severity;
+import org.sonar.api.rules.RuleType;
 import org.sonar.api.utils.DateUtils;
 import org.sonar.core.issue.DefaultIssue;
-import org.sonar.core.rule.RuleType;
 import org.sonar.db.issue.IssueDto;
 import org.sonar.db.issue.IssueMapper;
