+++ server/sonar-server/src/test/java/org/sonar/server/rule/index/RuleDocTesting.java
@@ -24,7 +24,7 @@ import java.util.Arrays;
 import org.sonar.api.rule.RuleKey;
 import org.sonar.api.rule.RuleStatus;
 import org.sonar.api.rule.Severity;
-import org.sonar.core.rule.RuleType;
+import org.sonar.api.rules.RuleType;
 import org.sonar.db.rule.RuleTesting;
 
 public class RuleDocTesting {
