+++ server/sonar-server/src/test/java/org/sonar/server/rule/index/RuleIndexerTest.java
@@ -27,8 +27,8 @@ import org.junit.Test;
 import org.sonar.api.config.Settings;
 import org.sonar.api.rule.RuleStatus;
 import org.sonar.api.rule.Severity;
+import org.sonar.api.rules.RuleType;
 import org.sonar.api.utils.System2;
-import org.sonar.core.rule.RuleType;
 import org.sonar.db.DbClient;
 import org.sonar.db.DbSession;
 import org.sonar.db.DbTester;
