+++ sonar-core/src/main/java/org/sonar/core/issue/DefaultIssue.java
@@ -46,9 +46,9 @@ import org.sonar.api.issue.Issue;
 import org.sonar.api.issue.IssueComment;
 import org.sonar.api.rule.RuleKey;
 import org.sonar.api.rule.Severity;
+import org.sonar.api.rules.RuleType;
 import org.sonar.api.utils.Duration;
 import org.sonar.core.issue.tracking.Trackable;
-import org.sonar.core.rule.RuleType;
 
 import static java.lang.String.format;
