+++ sonar-db/src/main/java/org/sonar/db/version/v55/FeedRulesTypes.java
@@ -23,8 +23,8 @@ import com.google.common.base.Joiner;
 import com.google.common.base.Splitter;
 import java.sql.SQLException;
 import java.util.List;
+import org.sonar.api.rules.RuleType;
 import org.sonar.api.utils.System2;
-import org.sonar.core.rule.RuleType;
 import org.sonar.db.Database;
 import org.sonar.db.version.BaseDataChange;
 import org.sonar.db.version.MassUpdate;
