+++ sonar-plugin-api/src/test/java/org/sonar/api/rules/RuleTypeTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.core.rule;
+package org.sonar.api.rules;
 
 import org.junit.Rule;
 import org.junit.Test;
